A major rewrite of the Komi make file: * separated out the xml2lexc conversion in another makefile - this will make it easier to include it in the hfst build process * added proper dependency tracking of the xml files (so only one make command is needed) * general cleanup
I have NOT touched the dictionary making other than moving it to the new xml makefile. I have tested building fst's on my computer, but nothing beyond that. Everything is working fine for me, though.

This reworking is kind of premature, as it really should have been done as part of the new infra project. Consider this as a warm-up for the new infra.