Extended the exclusion list for the error model quite a bit: all digits, several symbols, and all non-cyrillic characters found in the transducer.
This should speed up the correction a bit, by making the error correction transducer and thus the search space for the suggestions smaller. The speed is still bad, but should not be as bad as earlier.

The alphabetical characters in this list is also a good source for looking for suspicious entries in the lexicon - why does greek I (iota) appear in the symbol table? What about all the latin characters? Please have a look.