The i and ö in the lexicon TESTINGTESTING all were from the Latin alphabet (the reason was that they were taken from a Windows-based source without Komi keyboard.
The lexicon is taken from MWE in Enye Lev's material,
they should probably be done in the following way:
MWEs shouhd have the first part tagged as +Attr. Nouns
that can be first part of compounds (all nouns?) shall
have +Attr as one of its analyses, this analysis shall
then be discarded in CG afterwards whenever the noun is
not followed by another noun. The argument for this is
the same as for cases like "Margaret Thatcheris lea...",
analysing "Margaret" as +Nom will be problematic for the
syntactic analysis; the same will hold for Komi N N 
compounds.

But let us look at the TESTINGTESTING material first.